+++ app/src/main/java/com/example/newapppp/domain/model/Habit.kt
@@ -13,6 +13,6 @@ data class Habit(
     val color: HabitColor,
     val priority: HabitPriority,
     val type: HabitType,
-    val doneDates: List<Int>,
+    val doneDates: List<Long>,
     val frequency: Int
 ) : Parcelable
