+++ app/src/main/java/com/example/newapppp/domain/model/HabitUI.kt
@@ -9,6 +9,6 @@ data class HabitUI(
     val color: HabitColor,
     val priority: HabitPriority,
     val type: HabitType,
-    val doneDates: List<Int>,
+    val doneDates: List<Long>,
     val frequency: Int
 )
